refactor: move constants from modals/index.ts to types/const.ts; this prevents modal dialogs from being pulled in as dependencies to library code
