+++ src/types/const.ts
@@ -21,6 +21,26 @@ export enum SelectionType {
     NONE = "NONE"
 }
 
+// Type of dialog being edited.
+export enum EditDialogType {
+    NEW = 'NEW',
+    TRAIN_EDITED = 'TRAIN_EDITED',
+    TRAIN_ORIGINAL = 'TRAIN_ORIGINAL',
+    LOG_EDITED = 'LOG_EDITED',
+    LOG_ORIGINAL = 'LOG_ORIGINAL',
+    BRANCH = 'BRANCH',
+    IMPORT = 'IMPORT'
+}
+
+// State of current edited dialog.
+export enum EditState {
+    CAN_EDIT = 'CAN_EDIT',
+    // Running bot not compatible with Model.
+    INVALID_BOT = 'INVALID_BOT',
+    // Attempting to edit older package id.
+    INVALID_PACKAGE = 'INVALID_PACKAGE'
+}
+
 export const CL_IMPORT_TUTORIALS_USER_ID = '9c110735ea8b440d8f31c5c68ffb767d'
 
 export const BOT_HOST_NAME: Readonly<string> = location.hostname
